src/test/java/com/leetcode/tree/easy/FindModeInBinarySearchTree_501Test.java
package com.leetcode.tree.easy;

import com.leetcode.utils.BinaryTree;
import com.leetcode.utils.BinaryTree.TreeNode;
import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class FindModeInBinarySearchTree_501Test {
  private final FindModeInBinarySearchTree_501 solution = new FindModeInBinarySearchTree_501();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(arguments(BinaryTree.of(1, null, 2, 2), new int[] {2}));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void findMode(TreeNode root, int[] expectedResult) {
    int[] actualResult = solution.findMode(root);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
